+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## NEXT
  - Move `neo4j::Address` to `neo4j::address::Address`
+ - Impl `FromStr` for `neo4j::driver::ConnectionConfig` (besides `TryFrom<&str>`).
  - Much more documentation
 
 ## 0.0.1